FAR patch 01639.Mix Дата              : 21.04.2003 Сделал            : Valentin Skirdin Описание          : Mix Измененные файлы  : checkver.cpp clipboard.cpp config.cpp copy.cpp ctrlobj.cpp                     delete.cpp dialog.cpp fileedit.cpp findfile.cpp fn.hpp                     global.cpp global.hpp interf.cpp keyboard.cpp main.cpp                     mix.cpp plugin.hpp plugins.cpp plugins.hpp struct.hpp                     syslog.cpp message.cpp Новые файлы       : Удаленные файлы   : Состав            : 01639.Mix.txt                     checkver.cpp.1639.diff        clipboard.cpp.1639.diff                     config.cpp.1639.diff          copy.cpp.1639.diff                     ctrlobj.cpp.1639.diff         delete.cpp.1639.diff                     dialog.cpp.1639.diff          fileedit.cpp.1639.diff                     findfile.cpp.1639.diff        fn.hpp.1639.diff                     global.cpp.1639.diff          global.hpp.1639.diff                     interf.cpp.1639.diff          keyboard.cpp.1639.diff                     main.cpp.1639.diff            mix.cpp.1639.diff                     plugin.hpp.1639.diff          plugins.cpp.1639.diff                     plugins.hpp.1639.diff         struct.hpp.1639.diff                     syslog.cpp.1639.diff          message.cpp.1639.diff Основан на патче  : 1638 Дополнение        :
1. Copy:
   В обработчике DN_KEY сделаем "break", а не "return". Логика не
   изменится, но суть поменяется!
   Плюс - попытка исправить багу за нумбером 853

2. /co преобладает над /p - как и написанов в хелпе!

3. DM_GETSELECTION
     Получить параметры выделения в строках ввода в диалогах
     Msg=DM_GETSELECTION,
     Param1=ID,
     Param2=*EditorSelect

   DM_SETSELECTION
     Установить выделение в строках ввода в диалогах
     Msg=DM_SETSELECTION,
     Param1=ID,
     Param2=*EditorSelect

   В EditorSelect пока что юзаются только BlockStartPos и BlockWidth
   (эта структура выбрана в связи с... - когда-нить я все же доделаю
   DI_MEMOEDIT :-))

   Причем DM_GETSELECTION заполняет структуру вот так:
   BlockType=BTYPE_STREAM;
   BlockStartLine=0;
   BlockHeight=1;

   2DJ: (без фич ну ни куда :-)) Если будет желание - намылю описалово,
        а пока... "внутренняя фича".


4. +USE_WFUNC_ALW
   Это для варианта с опцией /8 - не юзать Unicode-консоль

5. W-функции для входных событий (отладочный вариант)
   Мля... я сунулся сюда и мне стало страшно:

   "...- Львы, медведи и тигры, - пробормотал А.Беттик.
       - Они самые. Проникнув в мегасферу, мой отец спросил у
         Иск-Ина по имени Уммон, чего боится Техно-Центр. Тот
         ответил, что элементы Техно-Центра боятся метасферы,
         где кишмя кишат львы, медведи и тигры..."


6. DETECT_ALT_ENTER - для тестирования варианта по перехвату
   Alt-Enter с последующим корректным отображением экрана!

7. "...Если у макрокоманды стоит атрибут Run after FAR start, она выполняется
   на старте, но... всё что идёт до первого пробела "обрезается". Другими
   словами, когда она выглядит как
     s e t Space P A T H = ...
   то выдаваемая системе команда выглядит как
     PATH=...
   что, понятно, приводит к ошибке..."

   "...Помнишь я рассказывал про "пропадающее" первое слово (до первого
   пробела) в макрокоманде выполняемой на старте фара? Рассказываю как
   такой эффект получить под форточками (любыми). Положи в директорию
   плагинов, dll которая не может быть загружена (ну, например ссылается на
   другую dll, которой нет в пути, или возвращает 0 из DLL_PROCESS_ATTACH).
   Фар в этой ситуации выводит окошко 'plugin.... can't be loaded'.
   Ну а тут начинает выполняться макрокоманда и... первый пробел из неё
   воспринимается как кнопка "закрывающая" это окно :).
   Соответственно всё что шло до этого пробела "съелось", а остальное
   попадает в командную строку..."

8. ClipBoard:
   Учтем состояние, когда AnsiMode=1 - в этом случае не "корежим"
   передаваемые в/из клипьорда данные (осталось докрутить редактор)
   Для этого: редактор должен вызывать InternalCopyToClipboard и
   InternalPasteFromClipboard/InternalPasteFromClipboardEx

   Это задел на будущее :-) По правде сказать я и редактор ковырял,
   но запутался и... оставил только эти функции.

9. Про _beginthread
   Уточним. Если трид не удалось создать, то функции вызываются напрамую,
   но если не предпренять спец.мер, то _endthread вываливат сам ФАР.
   Т.е. если RegistrationBugs==FALSE, то вызываем _endthread

10. "Трохе для сэбэ".
   Задолбал меня удалятор: для него THREAD_PRIORITY_NORMAL - это слишком
   жирно. Добавил (для себя любимого :-) кей в реестре
   "System\DelThreadPriority:dword" = одному из значений THREAD_PRIORITY_*
   Выставил не много ни мало - 0xffffffff - сразу полегчало.
   Если будет желание - можем оформить ЭТО в качестве TechInfo.
   По умолчанию значение кея оставил as THREAD_PRIORITY_NORMAL.

11.В удаляторе:
   Исключим лишние телодвижения, если LastError входит в диапазон
   ошибок, определенных в функции: CheckErrorForProcessed

     //Проверка на "продолжаемость" экспериментов по... например,
     //удалению файла с разными именами!
     BOOL CheckErrorForProcessed(DWORD Err)

   Это по поводу BugZ#857 - зря, всё-таки, фар не проверяет GetLastError:

      From: Юрий Харон <yjh@styx.cabel.net>

      Зря, всё-таки, фар не проверяет GetLastError :). Вот то на что я
      "нарвался" - DeleteFile (F8) возвращает ERROR_ACCESS_DENIED. Фар, "не
      глядючи" на код ошибки пытается повторить операцию сначала по полному
      имени, а потом (если ошибка осталась - а с чего бы такой ошибке
      исчезать? :) по имени вида "\\?\fullpathname". Ну в NT и Me/98 это
      приводит только к 4хкратному повторению операции (нафиг не нужному),
      а вот в каких-нито 95х/NT4 это приводит ещё и к неправильной
      диагностике - формат "\\?\" появился только в W2K, посему при попытке
      удаления залоченного файла, фар выводит FILE NAME INVALID (ответ
      последнего DeleteFile), что смотрится, согласись странновато :).

+++ far/delete.cpp
@@ -5,10 +5,14 @@ delete.cpp
 
 */
 
-/* Revision: 1.58 20.02.2003 $ */
+/* Revision: 1.59 21.04.2003 $ */
 
 /*
 Modify:
+  21.04.2003 SVS
+    + Opt.DelThreadPriority
+    ! Исключим лишние телодвижения, если LastError входит в диапазон
+      ошибок, определенных в функции: CheckErrorForProcessed
   20.02.2003 SVS
     ! Заменим strcmp(FooBar,"..") на TestParentFolderName(FooBar)
   10.02.2003 SVS
@@ -182,7 +186,7 @@ static int WipeDirectory(const char *Name);
 static void PR_ShellDeleteMsg(void);
 
 static int ReadOnlyDeleteMode,SkipMode,SkipFoldersMode,DeleteAllFolders;
-
+static int localLastError;
 static clock_t DeleteStartTime;
 
 enum {DELETE_SUCCESS,DELETE_YES,DELETE_SKIP,DELETE_CANCEL};
@@ -200,17 +204,21 @@ BOOL WINAPI FAR_DeleteFile(const char *FileName)
   BOOL rc=DeleteFile(FileName);
   if(!rc) // IS: вот тут лишние телодвижения и начнем...
   {
-    char FullName[NM*2]="\\\\?\\";
-    // IS: +4 - чтобы не затереть наши "\\?\"
-    if(ConvertNameToFull(FileName, FullName+4, sizeof(FullName)-4) < sizeof(FullName)-4)
+    SetLastError((localLastError = GetLastError()));
+    if(CheckErrorForProcessed(localLastError))
     {
-      // IS: проверим, а вдруг уже есть есть нужные символы в пути
-      if( (FullName[4]=='/' && FullName[5]=='/') ||
-          (FullName[4]=='\\' && FullName[5]=='\\') )
-        rc=DeleteFile(FullName+4);
-      // IS: нужных символов в пути нет, поэтому используем наши
-      else
-        rc=DeleteFile(FullName);
+      char FullName[NM*2]="\\\\?\\";
+      // IS: +4 - чтобы не затереть наши "\\?\"
+      if(ConvertNameToFull(FileName, FullName+4, sizeof(FullName)-4) < sizeof(FullName)-4)
+      {
+        // IS: проверим, а вдруг уже есть есть нужные символы в пути
+        if( (FullName[4]=='/' && FullName[5]=='/') ||
+            (FullName[4]=='\\' && FullName[5]=='\\') )
+          rc=DeleteFile(FullName+4);
+        // IS: нужных символов в пути нет, поэтому используем наши
+        else
+          rc=DeleteFile(FullName);
+      }
     }
   }
   return rc;
@@ -224,17 +232,21 @@ BOOL WINAPI FAR_RemoveDirectory(const char *DirName)
   BOOL rc=RemoveDirectory(DirName);
   if(!rc) // IS: вот тут лишние телодвижения и начнем...
   {
-    char FullName[NM+16]="\\\\?\\";
-    // IS: +4 - чтобы не затереть наши "\\?\"
-    if(ConvertNameToFull(DirName, FullName+4, sizeof(FullName)-4) < sizeof(FullName)-4)
+    SetLastError((localLastError = GetLastError()));
+    if(CheckErrorForProcessed(localLastError))
     {
-      // IS: проверим, а вдруг уже есть есть нужные символы в пути
-      if( (FullName[4]=='/' && FullName[5]=='/') ||
-          (FullName[4]=='\\' && FullName[5]=='\\') )
-        rc=RemoveDirectory(FullName+4);
-      // IS: нужных символов в пути нет, поэтому используем наши
-      else
-        rc=RemoveDirectory(FullName);
+      char FullName[NM+16]="\\\\?\\";
+      // IS: +4 - чтобы не затереть наши "\\?\"
+      if(ConvertNameToFull(DirName, FullName+4, sizeof(FullName)-4) < sizeof(FullName)-4)
+      {
+        // IS: проверим, а вдруг уже есть есть нужные символы в пути
+        if( (FullName[4]=='/' && FullName[5]=='/') ||
+            (FullName[4]=='\\' && FullName[5]=='\\') )
+          rc=RemoveDirectory(FullName+4);
+        // IS: нужных символов в пути нет, поэтому используем наши
+        else
+          rc=RemoveDirectory(FullName);
+      }
     }
   }
   return rc;
@@ -243,7 +255,7 @@ BOOL WINAPI FAR_RemoveDirectory(const char *DirName)
 
 void ShellDelete(Panel *SrcPanel,int Wipe)
 {
-  ChangePriority ChPriority(THREAD_PRIORITY_NORMAL);
+  ChangePriority ChPriority(Opt.DelThreadPriority);
   WIN32_FIND_DATA FindData;
   char DeleteFilesMsg[300],SelName[NM],SelShortName[NM],DizName[NM];
   char FullName[2058];
@@ -770,8 +782,7 @@ int ShellRemoveFile(const char *Name,const char *ShortName,int Wipe)
         break;
     }
     else
-      if (WinVer.dwPlatformId==VER_PLATFORM_WIN32_NT && WinVer.dwMajorVersion<4 ||
-          !Opt.DeleteToRecycleBin)
+      if (WinVer.dwPlatformId==VER_PLATFORM_WIN32_NT && WinVer.dwMajorVersion<4 || !Opt.DeleteToRecycleBin)
       {
 /*
         if (WinVer.dwPlatformId==VER_PLATFORM_WIN32_NT)
@@ -840,11 +851,11 @@ int ERemoveDirectory(const char *Name,const char *ShortName,int Wipe)
   {
     if (Wipe)
     {
-      if (WipeDirectory(Name) || WipeDirectory(ShortName))
+      if (WipeDirectory(Name) || (localLastError != ERROR_ACCESS_DENIED && WipeDirectory(ShortName)))
         break;
     }
     else
-      if (FAR_RemoveDirectory(Name) || FAR_RemoveDirectory(ShortName))
+      if (FAR_RemoveDirectory(Name) || (localLastError != ERROR_ACCESS_DENIED && FAR_RemoveDirectory(ShortName)))
         break;
     /* $ 19.01.2003 KM
        Добавлен Skip и Skip all
@@ -996,28 +1007,43 @@ int WipeFile(const char *Name)
   SetFilePointer(WipeHandle,0,NULL,FILE_BEGIN);
   SetEndOfFile(WipeHandle);
   CloseHandle(WipeHandle);
+
   char TempName[NM];
-  MoveFile(Name,FarMkTempEx(TempName,NULL,FALSE));
-  return(FAR_DeleteFile(TempName));
+  if(MoveFile(Name,FarMkTempEx(TempName,NULL,FALSE)))
+    return(FAR_DeleteFile(TempName));
+  SetLastError((localLastError = GetLastError()));
+  return FALSE;
 }
 
 
 int WipeDirectory(const char *Name)
 {
   char TempName[NM];
-  MoveFile(Name,FarMkTempEx(TempName,NULL,FALSE));
+  BOOL Ret=MoveFile(Name,FarMkTempEx(TempName,NULL,FALSE));
+  if(!Ret)
+  {
+    SetLastError((localLastError = GetLastError()));
+    return FALSE;
+  }
   return(FAR_RemoveDirectory(TempName));
 }
 
 int DeleteFileWithFolder(const char *FileName)
 {
   char FolderName[NM],*Slash;
-  SetFileAttributes(FileName,0);
-  remove(FileName);
-  strncpy(FolderName,FileName,sizeof(FolderName)-1);
-  if ((Slash=strrchr(FolderName,'\\'))!=NULL)
-    *Slash=0;
-  return(FAR_RemoveDirectory(FolderName));
+  BOOL Ret=SetFileAttributes(FileName,0);
+  if(Ret)
+  {
+    if(!remove(FileName))
+    {
+      strncpy(FolderName,FileName,sizeof(FolderName)-1);
+      if ((Slash=strrchr(FolderName,'\\'))!=NULL)
+        *Slash=0;
+      return(FAR_RemoveDirectory(FolderName));
+    }
+  }
+  SetLastError((localLastError = GetLastError()));
+  return FALSE;
 }
